requirements.txt
pandas
tensorflow
Pillow
numpy
tqdm
